# Significant speedup of code gen