## Summary

<!-- Describe your changes in detail here, if it closes an open issue, include "Closes #<issue>" -->
Replaced the map with a switch case as an experiment, turns out it's significantly faster so I'm keeping it!

```plaintext
❯ benchstat before.txt after.txt
goos: darwin
goarch: arm64
pkg: github.com/FollowTheProcess/hue
cpu: Apple M1 Pro
                  │ before.txt  │              after.txt              │
                  │   sec/op    │   sec/op     vs base                │
Style/simple-8      6.529n ± 0%   2.067n ± 0%  -68.33% (p=0.000 n=30)
Style/composite-8   90.31n ± 0%   71.48n ± 0%  -20.85% (p=0.000 n=30)
geomean             24.28n        12.16n       -49.94%

                  │  before.txt  │              after.txt              │
                  │     B/op     │    B/op     vs base                 │
Style/simple-8      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
Style/composite-8   16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=30) ¹
geomean                        ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                  │  before.txt  │              after.txt              │
                  │  allocs/op   │ allocs/op   vs base                 │
Style/simple-8      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
Style/composite-8   1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=30) ¹
geomean                        ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```